+++ src/classes/net/jankenpoi/sudokuki/model/GridChangedEvent.java
@@ -38,7 +38,7 @@ public class GridChangedEvent extends EventObject {
 	}
 
 	public short getInfos() {
-		return (short)infos;
+		return infos;
 	}
 	
 }
